Backend/requirements.txt
flask==2.3.2
textblob==0.17.1
flask-cors==3.0.10
pymongo==4.5.0
gunicorn==20.1.0
python-dotenv==1.0.0